+++ dotfiles.powershell/Microsoft.PowerShellISE_profile.allUsers.ps1
@@ -1,0 +1,5 @@
+#Requires -Version 5
+
+# Profile name              Location
+# ------------              --------
+# AllUsersISEHost           $psHome\Microsoft.PowerShellISE_profile.ps1