+++ dotfiles.powershell/Microsoft.PowerShellISE_profile.ps1
@@ -1,0 +1,13 @@
+#Requires -Version 5
+
+# Profile name              Location
+# ------------              --------
+# CurrentUserISEHost        $Home\Documents\WindowsPowerShell\Microsoft.PowerShellISE_profile.ps1
+
+
+
+
+# Initialization code
+
+    # Powershell window title
+        $host.UI.RawUI.WindowTitle += "   |   $($host.Name)"