+++ dotfiles.powershell/Microsoft.PowerShell_profile.allUsers.ps1
@@ -1,0 +1,24 @@
+﻿#Requires -Version 5
+
+# Profile name              Location
+# ------------              --------
+# AllUsersConsoleHost       $psHome\Microsoft.PowerShell_profile.ps1
+
+
+
+
+# Initialization code
+
+    # Keep track of all profile scripts launched by Powershell during session initialization
+    $global:profiles += , $psCommandPath
+
+
+
+
+
+
+
+
+#region Cleanup
+
+#endregion