+++ dotfiles.powershell/profile/Public/Invoke-DefaultProfile.ps1
@@ -1,5 +1,0 @@
-function Invoke-DefaultProfile {
-
-    return $True
-    
-}