+++ Release-Notes.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## v0.1.3
+
+Change profile scripts directory structure;
+Public functions injection completed;
+Support of All Users & Current User Powershell profiles;
+Support of Console Host, ISE Host and All Hosts versions of Powershell profile;
+Note: external dependency on sudo/Invoke-ElevatedCommand is introduced, TBD
+
 ## v0.1.2
 
 Add Invoke-PSDeploy for deployment of artifacts; Change folder layout